Move login methods to auth
